Correction to r6160
git-svn-id: http://svn.osgeo.org/qgis/trunk@6191 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/qgis.h
@@ -100,7 +100,7 @@ public:
     ProviderCountCalcEvent
   };
   
-  static const int DEFAULT_IDENTIFY_RADIUS;
+  static const double DEFAULT_IDENTIFY_RADIUS;
 };
   /** WKT string that represents a geographic coord sys */
   const  QString GEOWKT =
